+++ db/migrate/20140816174246_add_custom_layout.rb
@@ -1,0 +1,8 @@
+class AddCustomLayout < ActiveRecord::Migration
+  def change
+    add_column :users, :css_form_background, :string
+    add_column :users, :css_form_color, :string
+    add_column :users, :form_attachment, :boolean, default: true
+    add_column :users, :form_advanced_mode, :boolean, default: true
+  end
+end