+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20140812184133) do
+ActiveRecord::Schema.define(:version => 20140816174246) do
 
   create_table "messages", :force => true do |t|
     t.datetime "created_at", :null => false
@@ -23,20 +23,24 @@ ActiveRecord::Schema.define(:version => 20140812184133) do
   create_table "users", :force => true do |t|
     t.string   "username"
     t.text     "public_key"
-    t.datetime "created_at",                             :null => false
-    t.datetime "updated_at",                             :null => false
-    t.string   "email",                  :default => "", :null => false
-    t.string   "encrypted_password",     :default => "", :null => false
+    t.datetime "created_at",                               :null => false
+    t.datetime "updated_at",                               :null => false
+    t.string   "email",                  :default => "",   :null => false
+    t.string   "encrypted_password",     :default => "",   :null => false
     t.string   "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          :default => 0,  :null => false
+    t.integer  "sign_in_count",          :default => 0,    :null => false
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
     t.string   "last_sign_in_ip"
     t.string   "customer_id"
     t.string   "last_4_digits"
+    t.string   "css_form_background"
+    t.string   "css_form_color"
+    t.boolean  "form_attachment",        :default => true
+    t.boolean  "form_advanced_mode",     :default => true
   end
 
   add_index "users", ["email"], :name => "index_users_on_email", :unique => true