+++ Changelog.md
@@ -1,3 +1,15 @@
+# Next
+
+
+## Features
+
+* Added embed link
+* Added custom form layout
+
+## Bug fixes
+
+## Refactor
+
 # 0.3
 
 Aug 15,2014